mv rthw api to separate file
