+++ components/drivers/include/drivers/pic.h
@@ -122,7 +122,7 @@ struct rt_pic_irq
     struct rt_pic_irq *parent;
 };
 
-rt_err_t rt_pic_linear_irq(struct rt_pic *pic, rt_size_t irq_nr);
+void rt_pic_default_name(struct rt_pic *pic);
 struct rt_pic *rt_pic_dynamic_cast(void *ptr);
 
 rt_err_t rt_pic_linear_irq(struct rt_pic *pic, rt_size_t irq_nr);
@@ -141,6 +141,7 @@ rt_inline struct rt_pic_irq *rt_pic_find_irq(struct rt_pic *pic, int irq_index)
 }
 
 struct rt_pic_irq *rt_pic_find_ipi(struct rt_pic *pic, int ipi_index);
+struct rt_pic_irq *rt_pic_find_pirq(struct rt_pic *pic, int irq);
 
 rt_err_t rt_pic_cascade(struct rt_pic_irq *pirq, int parent_irq);
 rt_err_t rt_pic_uncascade(struct rt_pic_irq *pirq);
@@ -169,7 +170,7 @@ rt_err_t rt_pic_irq_set_affinity(int irq, rt_bitmap_t *affinity);
 rt_err_t rt_pic_irq_get_affinity(int irq, rt_bitmap_t *out_affinity);
 rt_err_t rt_pic_irq_set_triger_mode(int irq, rt_uint32_t mode);
 rt_uint32_t rt_pic_irq_get_triger_mode(int irq);
-void rt_pic_irq_send_ipi(int irq, unsigned int cpumask);
+void rt_pic_irq_send_ipi(int irq, rt_bitmap_t *cpumask);
 
 void rt_pic_irq_parent_enable(struct rt_pic_irq *pirq);
 void rt_pic_irq_parent_disable(struct rt_pic_irq *pirq);
