+++ components/drivers/pic/SConscript
@@ -8,7 +8,7 @@ if not GetDepend(['RT_USING_PIC']):
 cwd = GetCurrentDir()
 CPPPATH = [cwd + '/../include']
 
-src = ['pic.c']
+src = ['pic.c', 'pic_rthw.c']
 
 if GetDepend(['RT_PIC_ARM_GIC']) or GetDepend(['RT_PIC_ARM_GIC_V3']):
     src += ['pic-gic-common.c']
