+++ components/drivers/pic/pic-gicv3.h
@@ -124,12 +124,12 @@
 #define GICR_TYPER_DirectLPIS   (1U << 3)
 #define GICR_TYPER_LAST         (1U << 4)
 #define GICR_TYPER_RVPEID       (1U << 7)
-#define GICR_TYPER_COM_LPI_AFF  GENMASK_ULL(25, 24)
-#define GICR_TYPER_AFFINITY     GENMASK_ULL(63, 32)
+#define GICR_TYPER_COM_LPI_AFF  RT_GENMASK_ULL(25, 24)
+#define GICR_TYPER_AFFINITY     RT_GENMASK_ULL(63, 32)
 
-#define GICR_INVLPIR_INTID      GENMASK_ULL(31, 0)
-#define GICR_INVLPIR_VPEID      GENMASK_ULL(47, 32)
-#define GICR_INVLPIR_V          GENMASK_ULL(63, 63)
+#define GICR_INVLPIR_INTID      RT_GENMASK_ULL(31, 0)
+#define GICR_INVLPIR_VPEID      RT_GENMASK_ULL(47, 32)
+#define GICR_INVLPIR_V          RT_GENMASK_ULL(63, 63)
 
 #define GICR_INVALLR_VPEID      GICR_INVLPIR_VPEID
 #define GICR_INVALLR_V          GICR_INVLPIR_V
@@ -145,8 +145,8 @@
 #define GICR_WAKER_ChildrenAsleep   (1U << 2)
 
 #define GICR_PROPBASER_IDBITS_MASK  (0x1f)
-#define GICR_PROPBASER_ADDRESS(x)   ((x) & GENMASK_ULL(51, 12))
-#define GICR_PENDBASER_ADDRESS(x)   ((x) & GENMASK_ULL(51, 16))
+#define GICR_PROPBASER_ADDRESS(x)   ((x) & RT_GENMASK_ULL(51, 12))
+#define GICR_PENDBASER_ADDRESS(x)   ((x) & RT_GENMASK_ULL(51, 16))
 
 /* ITS registers */
 #define GITS_CTLR               0x0000
